Creating Web Client v0.5
A lot of code goes into the initial commit.

The web client has an admin panel to test the Parite Framework, as well as a first version of the game host screen, with UX for the whole table, cards, players, etc.

The code still contains some React playground unrelated code as well.